refactor(US-005): collapse execution.ts to wrapper-only inspection (AC#5)
- Extract applyPostRunInspection and decideStageAction to src/execution/post-run.ts
- Extract extractPauseReason and deriveTddFailureCategory to post-run.ts
- Add assemblePlanInputsFromCtx to src/execution/plan-inputs.ts
- Collapse execution.ts from 512 lines to 130 lines (wrapper-only)
- Export new symbols from src/execution/index.ts barrel
- All 2472 tests pass, typecheck clean
